Cast is_visible to integer before binding

+++ backend/app/Models/CollegeUniversityManagementModel.php
@@ -82,7 +82,7 @@ class CollegeUniversityManagementModel
 
         $stmt = $this->pdo->prepare($query);
 
-        $is_visible = strip_tags($data['is_visible']);
+        $is_visible = (int) strip_tags($data['is_visible']);
 
         $stmt->bindParam(':is_visible', $is_visible);
         $stmt->bindParam(':id', $id);
